+++ implementations/hotchocolate/Types/ProductDimension.cs
