+++ implementations/hotchocolate/Types/ProductVariation.cs
